+++ apps/backend/src/Obscura.Contracts/Entities/Capabilities/Core/EntityCapability.cs
@@ -12,7 +12,6 @@ namespace Obscura.Contracts.Entities;
 [JsonDerivedType(typeof(LinksCapability), "links")]
 [JsonDerivedType(typeof(FlagsCapability), "flags")]
 [JsonDerivedType(typeof(FilesCapability), "files")]
-[JsonDerivedType(typeof(CountersCapability), "counters")]
 [JsonDerivedType(typeof(FingerprintsCapability), "fingerprints")]
 [JsonDerivedType(typeof(MarkersCapability), "markers")]
 [JsonDerivedType(typeof(SubtitlesCapability), "subtitles")]